testing framework for quasi-likelihood problems
