+++ src/problems/ql_logistic_sin.jl
@@ -80,7 +80,7 @@ Construct a quasi-likelihood estimation problem with
     The `design` matrix has a column of all ones and the rest generated from a 
     normal distribution with its entries divided by `sqrt(nvar - 1)`. 
     The `response` vector has entries generated by 
-    `mean.(design * x) + variance.(mean.(design * x)) * ϵ`, where `ϵ` is a noise
+    `mean.(design * x) + variance.(mean.(design * x)) .^ (.5) * ϵ`, where `ϵ` is a noise
     vector generated from the Arcsine distribution with default parameters.
  
     function QLLogisticSin(design::Matrix{T}, response::Vector{T}, 
@@ -145,7 +145,7 @@ function QLLogisticSin(
     ϵ = T.((rand(Distributions.Arcsine()) .- .5)./(1/8)) # standardize
 
     # generate responses
-    response = μ_obs + OptimizationMethods.linear_plus_sin.(μ_obs) * ϵ
+    response = μ_obs + T.(OptimizationMethods.linear_plus_sin.(μ_obs) .^ (.5)) * ϵ
 
     return QLLogisticSin{T, Vector{T}}(
         meta,
